+++ src/lib/constants/layout.ts
@@ -1,4 +1,5 @@
 export const BANNER_HEIGHT = 63.5;
 export const PLAYER_HEIGHT = 80;
 export const MENU_WIDTH = 240;
-export const PRIVATE_MENU_ITEM_BG_COLOR = "#22c55e";
+export const PRIVATE_MENU_ITEM_BG_COLOR = "#ffffff";
+export const SPOTIFY_MENU_ITEM_BG_COLOR = "#1db954";